+++ service-management/src/main/java/com/ummbatin/service_management/repositories/ComplaintRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ComplaintRepository extends JpaRepository<Complaint, Integer> {
-    List<Complaint> findByResident_ResidentId(Integer residentId);
+public interface ComplaintRepository extends JpaRepository<Complaint, Long> {
+    List<Complaint> findByResident_ResidentId(Long residentId);
 }
